+++ Modules/Microsoft365DSC/Modules/M365DSCDRGUtil.psm1
@@ -1298,7 +1298,11 @@ function Update-DeviceConfigurationPolicyAssignment
         [Parameter()]
         [ValidateSet('v1.0','beta')]
         [System.String]
-        $APIVersion = 'beta'
+        $APIVersion = 'beta',
+
+        [Parameter()]
+        [System.String]
+        $RootIdentifier = 'assignments'
     )
 
     try
@@ -1331,8 +1335,10 @@ function Update-DeviceConfigurationPolicyAssignment
             }
             $deviceManagementPolicyAssignments += @{'target' = $formattedTarget}
         }
-        $body = @{'assignments' = $deviceManagementPolicyAssignments} | ConvertTo-Json -Depth 20
-        write-verbose -Message $body
+
+        $body = @{$RootIdentifier = $deviceManagementPolicyAssignments} | ConvertTo-Json -Depth 20
+        Write-Verbose -Message $body
+        
         Invoke-MgGraphRequest -Method POST -Uri $Uri -Body $body -ErrorAction Stop
     }
     catch